WIP moving away from "api" to "websockets".
This is somewhat related to [refs #157171346] and a slew of other
races. Currently, we avoid racing by serializing requests on the client.
But we can be way faster if we just use Websockets -- which gives us
the serialization guarantees we want, _and_ reduces server overhead,
_and_ gives server affinity, _and_ doesn't lock us in to a public API.

This is inspired by Tabs, which break our entire API. It shouldn't be
a public API yet. It should be this.

The new framework is primarily async. Be wary when accessing the database.